requirements.txt
boto3==1.26.120
botocore==1.29.120
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fillpdf==0.7.2
Flask==2.2.3
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
pdf2image==1.16.2
pdfrw2==0.5.0
Pillow==9.4.0
pycryptodome==3.17
PyMuPDF==1.21.1
pypdf==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
reportlab==3.6.12
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
XlsxWriter==3.0.8